Temporary fix for a stupid bug. See full commit message for the full details.
Bug:
----
When the caret is on a list item, somewhere in the middle of the line,
and the user presses `<CR>`, then a new list item is added on a new line
instead of splitting the current line. This commit partly fixes that
bug.

Caveat:
-------
When the caret is on a list item, at the character previous to the last
one (`col('$') - 1`), and the user presses `<CR>`, then a new list item
will be added on a new line instead of splitting the current line.