Fix emphasis regions and update Rule match
This commit attempts to fix a problem with the emphasis regions interferring with
the horizontal rule match.  It also adds the missing underscore horizonal rule
syntax and consolidates the rule patterns.  See #86

I've left the `{1}` stuff in there and aligned the rules for readability.
Backslashitis makes my head hurt.  I can remove those things and rebase if I
need to.