Code block region regex improved.
At the moment, any line starting with 4 spaces or a tab is considered as
a code block, which causes issues in code like this:

	1. Agenda Item 1: Frob the Bazzit
	   - The bazzit is something that makes us have
		 very _short_ lines indeed.
	   - Further frobbing is necessary

The line "very _short_ lines indeed." is considered as a code block (and
therefore, the _short_ is not italised.

The change of this commit enforces a code block to be preceeded with an
empty line. This gives the following behaviour:

	1. Agenda Item 1: Frob the Bazzit
	   - The bazzit is something that makes us have
		 This is **NOT** a code block
	   - Further frobbing is necessary

		This **IS** a code block

		This **IS** also a code block
		And so is that

	Some text

		This **IS** also a code block
		Which is followed directly by some not code after this line.
	Some more text
	very _short_ lines indeed.
		This is **NOT** a code block

	very _short_ lines indeed.

For context, this comes from https://stackoverflow.com/q/55645317/3866623